# util/mpsc: refactor channel