--- a/Cargo.toml
+++ b/Cargo.toml
@@ -88,8 +88,7 @@ fail = "0.2"
 uuid = { version = "0.6", features = [ "serde", "v4" ] }
 grpcio = { version = "0.4", features = [ "secure" ] }
 raft = "0.4"
-crossbeam-channel = "0.2"
-crossbeam = "0.2"
+crossbeam = "0.5"
 fxhash = "0.2"
 derive_more = "0.11.0"
 num = "0.2"