--- a/benches/misc/channel/bench_channel.rs
+++ b/benches/misc/channel/bench_channel.rs
@@ -15,7 +15,7 @@ use std::sync::mpsc::channel;
 use std::thread;
 use test::Bencher;
 
-use crossbeam_channel;
+use crossbeam::channel;
 use mio::{EventLoop, Handler, Sender};
 use tikv::util::mpsc;
 
@@ -123,7 +123,7 @@ fn bench_util_channel(b: &mut Bencher) {
 
 #[bench]
 fn bench_util_loose(b: &mut Bencher) {
-    let (mut tx, rx) = mpsc::loose_bounded(480000);
+    let (tx, rx) = mpsc::loose_bounded(480000);
 
     let t = thread::spawn(move || {
         let mut n2: usize = 0;
@@ -150,7 +150,7 @@ fn bench_util_loose(b: &mut Bencher) {
 
 #[bench]
 fn bench_crossbeam_channel(b: &mut Bencher) {
-    let (tx, rx) = crossbeam_channel::unbounded();
+    let (tx, rx) = channel::unbounded();
 
     let t = thread::spawn(move || {
         let mut n2: usize = 0;
@@ -166,10 +166,10 @@ fn bench_crossbeam_channel(b: &mut Bencher) {
     let mut n1 = 0;
     b.iter(|| {
         n1 += 1;
-        tx.send(1)
+        tx.send(1).unwrap();
     });
 
-    tx.send(0);
+    tx.send(0).unwrap();
     let n2 = t.join().unwrap();
     assert_eq!(n1, n2);
 }