--- a/benches/misc/mod.rs
+++ b/benches/misc/mod.rs
@@ -15,7 +15,7 @@
 
 extern crate arrow;
 extern crate byteorder;
-extern crate crossbeam_channel;
+extern crate crossbeam;
 extern crate kvproto;
 extern crate log;
 extern crate mio;