--- a/src/lib.rs
+++ b/src/lib.rs
@@ -41,8 +41,6 @@ extern crate chrono;
 extern crate chrono_tz;
 extern crate crc;
 extern crate crossbeam;
-#[macro_use]
-extern crate crossbeam_channel;
 extern crate crypto;
 #[macro_use]
 extern crate fail;