- remove future support to make performance better
- add half close support
- add several useful APIs

Previous:

```
test channel::bench_channel::bench_crossbeam_channel                  ... bench:          24 ns/iter (+/- 0)
test channel::bench_channel::bench_util_channel                       ... bench:          75 ns/iter (+/- 1)
test channel::bench_channel::bench_util_loose                         ... bench:          90 ns/iter (+/- 5)
```

After:

```
test channel::bench_channel::bench_crossbeam_channel                  ... bench:          27 ns/iter (+/- 0)
test channel::bench_channel::bench_util_channel                       ... bench:          27 ns/iter (+/- 0)
test channel::bench_channel::bench_util_loose                         ... bench:          76 ns/iter (+/- 1)


```